mapqa: mechanism that allows us to parameterize search algorithms
MapQa is a trait that generalizes maps indexed by Qa, implemented for
Grids, HashMaps and BTreeMaps. This trait allows us to parameterize
the search algorithms so that we can use HashMaps/BTreeMaps when the
Qa dimensions are too large for Grids.

The first algorithms to support this parametrization is UCS.